+++ packages/connect/src/types/api/bitcoin/index.ts
@@ -1,5 +1,8 @@
 import type { PROTO } from '../../../constants';
 import type { AccountAddresses } from '@trezor/blockchain-link';
+import type { Transaction as BlockbookTransaction } from '@trezor/blockchain-link-types/lib/blockbook';
+
+import { AccountUtxo } from '../../account';
 
 // signMessage
 
@@ -66,6 +69,7 @@ export interface SignTransaction {
     refTxs?: RefTransaction[];
     account?: {
         addresses: AccountAddresses;
+        utxo: AccountUtxo[];
     };
     coin: string;
     locktime?: number;
@@ -89,6 +93,7 @@ export type SignedTransaction = {
     serializedTx: string;
     witnesses?: (string | undefined)[];
     txid?: string;
+    signedTransaction?: BlockbookTransaction;
 };
 
 // verifyMessage