+++ packages/suite/src/hooks/wallet/__fixtures__/useSendForm.ts
@@ -1364,6 +1364,11 @@ export const signAndPush = [
                 success: true,
                 payload: {
                     serializedTx: 'serializedABCD',
+                    signedTransaction: {
+                        txid: 'txid',
+                        vin: [],
+                        vout: [],
+                    },
                 },
             },
         ],